feat udistrital/arka_cliente#885: Id Cuentas
- Asiento contable en helper especifico
