+++ helpers/catalogoElementosHelper/catalogoElementosCrud.go
@@ -57,3 +57,19 @@ func GetAllDetalleSubgrupo(query string) (detalle []*models.DetalleSubgrupo, out
 
 	return detalle, nil
 }
+
+// GetSubgrupoById Consulta controlador subgrupo/{id} del api catalogo_elementos_crud
+func GetSubgrupoById(id int) (subgrupo *models.Subgrupo, outputError map[string]interface{}) {
+
+	funcion := "GetSubgrupoById"
+	defer errorctrl.ErrorControlFunction(funcion+" - Unhandled Error!", "500")
+
+	urlcrud := "http://" + beego.AppConfig.String("catalogoElementosService") + "subgrupo/" + strconv.Itoa(id)
+	if err := request.GetJson(urlcrud, &subgrupo); err != nil {
+		logs.Error(err)
+		eval := " - request.GetJson(urlcrud, &subgrupo)"
+		return nil, errorctrl.Error(funcion+eval, err, "500")
+	}
+
+	return subgrupo, nil
+}
